fix return await, ref, msgvalidator and naming. Also now user can get/edit/delete only his own messages

+++ src/controllers/message.ts
@@ -1,22 +1,25 @@
 import { Body, Controller, Ctx, Delete, Flow, Get, Params, Post } from 'amala'
 import { Context } from 'koa'
-import { MessageId, MessageText } from '@/validators/MessageValidators'
+import { MessageId, MessageText } from '@/validators/Message'
 import { MessageModel } from '@/models/message'
-import { forbidden, notFound } from '@hapi/boom'
-import attachUser from '@/middleware/attachUser'
+import { notFound } from '@hapi/boom'
+import authorize from '@/middleware/authorize'
 
 @Controller('/messages')
-@Flow(attachUser)
+@Flow(authorize)
 export default class MessageController {
   @Get('/')
-  async getAllMessages() {
-    return await MessageModel.find()
+  getAllMessages(@Ctx() ctx: Context) {
+    return MessageModel.find({ author: ctx.state.user._id })
   }
 
   @Get('/:id')
   async getMessage(@Ctx() ctx: Context, @Params() { id }: MessageId) {
     const message = await MessageModel.findById(id)
-    if (!message) {
+    if (
+      !message ||
+      ctx.state.user._id.toString() != message?.author?.toString()
+    ) {
       return ctx.throw(notFound("Can't find message with this id"))
     }
 
@@ -29,7 +32,7 @@ export default class MessageController {
     @Body({ required: true }) { text }: MessageText
   ) {
     const message = await MessageModel.create({
-      userId: ctx.state.user._id,
+      author: ctx.state.user._id,
       text,
     })
     return message
@@ -42,15 +45,11 @@ export default class MessageController {
     @Body({ required: true }) { text }: MessageText
   ) {
     const message = await MessageModel.findById(id)
-    if (!message) {
-      return ctx.throw(notFound("This message doesn't exist"))
-    }
-    if (ctx.state.user._id.toString() != message?.userId?.toString()) {
-      return ctx.throw(
-        forbidden(
-          'You have no access to this message. Users can edit or delete only their own messages'
-        )
-      )
+    if (
+      !message ||
+      ctx.state.user._id.toString() != message?.author?.toString()
+    ) {
+      return ctx.throw(notFound("Can't find message with this id"))
     }
 
     message.text = text
@@ -61,15 +60,11 @@ export default class MessageController {
   @Delete('/:id')
   async deleteMsg(@Ctx() ctx: Context, @Params() { id }: MessageId) {
     const message = await MessageModel.findById(id)
-    if (!message) {
-      return ctx.throw(notFound("This message doesn't exist"))
-    }
-    if (ctx.state.user._id.toString() != message?.userId?.toString()) {
-      return ctx.throw(
-        forbidden(
-          'You have no access to this message. Users can edit or delete only their own messages'
-        )
-      )
+    if (
+      !message ||
+      ctx.state.user._id.toString() != message?.author?.toString()
+    ) {
+      return ctx.throw(notFound("Can't find message with this id"))
     }
 
     await MessageModel.findByIdAndRemove(id)
